third_party/WebKit/LayoutTests/editing/spelling/spellcheck-disable-enable.html

<!DOCTYPE html>
<html>
<head>
<script src="../../fast/js/resources/js-test-pre.js"></script>
</head>
<body onload="test(document.getElementById('destination'), document.getElementById('frame').contentWindow.document);">
<pre id="console"></pre>
<div id="container">
  <div id="destination" contentEditable></div>
  <iframe id="frame" src="data:text/html,<body contenteditable></body>"></iframe>
</div>

<script>
description("Spell check markers should be removed from the whole page when disabling spell checker but " +
"they should be restored in the focused editable if spell checker gets enabled. " +
"To test manually type something with mispellings in the above editable element and iframe and turn " +
"spell checker off - misspelling markers should disappear. Having the editable focused " +
"turn spell checker on again. Misspellings in the editable should be marked again.");

jsTestIsAsync = true;

var destination_elm = null;
var destination_elm_in_frame = null;

function test(destination, frame_doc)
{
    if (!window.internals)
    {
        document.getElementById("console").innerHTML = "Automatic testing impossible. Test manually.\n";
        return;
    }

    internals.settings.setAsynchronousSpellCheckingEnabled(false);
    internals.settings.setUnifiedTextCheckerEnabled(true);

    destination_elm = destination;
    destination_elm_in_frame = frame_doc.body;
    destination_elm.focus();
    document.execCommand("InsertText", false, "zz.");
    shouldBe('internals.markerCountForNode(destination_elm.childNodes[0], "spelling")', '1');
    destination_elm_in_frame.focus();
    frame_doc.execCommand("InsertText", false, "zz.");
    shouldBe('internals.markerCountForNode(destination_elm_in_frame.childNodes[0], "spelling")', '1');
    internals.setContinuousSpellCheckingEnabled(false);
    shouldBe('internals.markerCountForNode(destination_elm.childNodes[0], "spelling")', '0');
    shouldBe('internals.markerCountForNode(destination_elm_in_frame.childNodes[0], "spelling")', '0');
    destination_elm.focus();
    internals.setContinuousSpellCheckingEnabled(true);
    shouldBe('internals.markerCountForNode(destination_elm.childNodes[0], "spelling")', '1');

    finishJSTest();
}

</script>
<script src="../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
