third_party/WebKit/LayoutTests/editing/spelling/spellcheck-disable-enable-expected.txt

Spell check markers should be removed from the whole page when disabling spell checker but they should be restored in the focused editable if spell checker gets enabled. To test manually type something with mispellings in the above editable element and iframe and turn spell checker off - misspelling markers should disappear. Having the editable focused turn spell checker on again. Misspellings in the editable should be marked again.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
PASS internals.markerCountForNode(destination_elm.childNodes[0], "spelling") is 1
PASS internals.markerCountForNode(destination_elm_in_frame.childNodes[0], "spelling") is 1
PASS internals.markerCountForNode(destination_elm.childNodes[0], "spelling") is 0
PASS internals.markerCountForNode(destination_elm_in_frame.childNodes[0], "spelling") is 0
PASS internals.markerCountForNode(destination_elm.childNodes[0], "spelling") is 1
zz.
